stash WIP changes to validation pp call oneway for riscv support

+++ example/validation_ppcall_one_way/run_mq.sh
@@ -1,5 +1,19 @@
 
-rm -rfd build && \
-make && \
-mq.sh run -s odroidc4_1 -f build/loader.img -c "FINISHED"
+if [ "$1" = "odroidc4" ];
+then
+    rm -rfd build && \
+    make build_odroidc4 && \
+    mq.sh run -s odroidc4_1 -f build/loader.img -c "BENCHFINISHED" && exit 0
 
+fi
+
+if [ "$1" = "hifive" ];
+then
+    rm -rfd build && \
+    make build_hifive && \
+    make -C ../../../opensbi -j$(nproc) PLATFORM=generic FW_PAYLOAD_PATH=$(pwd)/build/loader.img PLATFORM_RISCV_XLEN=64 PLATFORM_RISCV_ISA=rv64imac PLATFORM_RISCV_ABI=lp64 O=$(pwd)/build LLVM=1 && \
+    mq.sh run -s hifive -f build/platform/generic/firmware/fw_payload.bin -c "BENCHFINISHED" && exit 0
+fi
+
+echo "unknown board"
+exit 1
